split nv.rs into smaller files
