+++ ch32-metapac-gen/res/src/nv/mod.rs
@@ -1,0 +1,28 @@
+//! Runtime descriptor overlay for non-volatile flash structures (Option Bytes, ESIG, ...).
+//!
+//! Discovers `NvStruct` metadata attached to the active chip's memory regions, then exposes
+//! a path-based API (`descriptor.entry[.field]`) for reading, writing, and validating buffers
+//! laid out per those descriptors.
+
+mod codec;
+mod descriptor;
+mod lifecycle;
+mod types;
+
+#[cfg(test)]
+mod tests;
+
+pub use descriptor::Descriptor;
+pub use types::{EncodeError, EncodeInput, Info, ValidationError, Value};
+
+pub fn decode<'a>(path: &str, buf: &'a [u8]) -> Option<Value<'a>> {
+    let (descriptor, rest) = path.split_once('.')?;
+    Descriptor::find(descriptor)?.decode(buf, rest)
+}
+
+pub fn encode(path: &str, buf: &mut [u8], input: EncodeInput) -> Result<(), EncodeError> {
+    let (descriptor, rest) = path.split_once('.').ok_or(EncodeError::NoSuchEntry)?;
+    Descriptor::find(descriptor)
+        .ok_or(EncodeError::NoSuchEntry)?
+        .encode(buf, rest, input)
+}
